requirements.txt
absl-py==2.0.0
aiofiles==23.2.1
aiogram==3.3.0
aiohttp==3.9.1
aiosignal==1.3.1
aiosqlite==0.19.0
annotated-types==0.5.0
APScheduler==3.10.4
astunparse==1.6.3
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.1.0
Babel==2.9.1
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.1.1
cycler==0.12.1
databases==0.8.0
emoji==2.6.0
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
flatbuffers==23.5.26
fonttools==4.43.1
frozenlist==1.4.0
gast==0.5.4
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.0
h5py==3.10.0
idna==3.4
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
keras==2.14.0
kiwisolver==1.4.5
libclang==16.0.6
lxml==4.9.3
magic-filter==1.0.12
Markdown==3.5
MarkupSafe==2.1.3
matplotlib==3.8.0
ml-dtypes==0.2.0
multidict==6.0.4
numpy==1.26.1
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.1
Pillow==9.5.0
plotly==5.17.0
protobuf==4.24.4
psycopg2-binary==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.5.3
pydantic_core==2.14.6
pygame==2.4.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.11.3
six==1.16.0
SQLAlchemy==1.4.51
tenacity==8.2.3
tensorboard==2.14.1
tensorboard-data-server==0.7.1
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-macos==2.14.0
tensorflow-metal==1.1.0
termcolor==2.3.0
tqdm==4.66.1
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.2
urllib3==2.0.3
Werkzeug==2.3.7
wrapt==1.14.1
yarl==1.9.2
